Send the results of checking the tracker to bot
The parameters of the chat interaction with the bot (first name, chat
id) are stored in the event variable. Also passed is the variable
credlist generated in the previous function. For each credential in the
list of credentials this function gets the outputs from S3 (date of last
update and screenshot of the timeline) and sends them back into the chat
with the bot